refactor(buffer): split buffer module into files
